ProductionCost is stored in db with each material rate and cost and quantity with respect to production request

+++ database/migrations/2017_11_28_143445_create_gate_passes_table.php
@@ -19,7 +19,7 @@ class CreateGatePassesTable extends Migration
             $table->string('contact_phone');
             $table->string('destination');
             $table->string('items');
-            $table->float('quantity', 8, 4);
+            $table->double('quantity', 15, 4);
             $table->text('remarks');
             $table->timestamps();
         });
